Add a more generic way to manage GOV.UK surveys
The current code (in `user-satisfaction-survey.js`) ties together both
the mechanism to present a survey for the user to fill out, and the
metadata about the survey itself.

Historically, there has only been a single survey, but this is about to
change.

Therefore, create a more "generic" mechanism to display surveys, where
the metadata about a survey is separated from the mechanism of selecting
the "Active" survey and displaying it.

Thus, we can add more small surveys -- with varying start and end times,
and varying frequencies -- by adding a single set of metadata to the
array in `surveys.js`

Furthermore, there are identifiable enhancement paths where we can make
surveys conditionally appear on different parts of the site -- a
requirement coming up in a few weeks time.

With @NeilvB

TODO: Figure out how cache expiry needs to work when this change
deploys. What's the impact of it?