Refactor transaction survey prevention
The correct place for this logic is inside the survey javascript where
other decisions are made with a bearing on if we should display the
survey.

During this refactor, we moved the `window.GOVUK` guard to surround what
appears to also be relevant code that should be protected by this guard.

+ @NeilvB